Safari-611.1.10/LayoutTests/http/tests/security/window-defineProperty-crossOrigin.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<iframe src="http://localhost:8000/security/resources/iframe-with-element.html"></iframe>
<script>
description("Test that calling Object.defineProperty() on a cross-origin Window throws.");
jsTestIsAsync = true;

onload = function() {
    shouldThrowErrorName("Object.defineProperty(frames[0], 'foo', { value: 1 })", "SecurityError");
    finishJSTest();
};
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
