Safari-609.1.11.1/LayoutTests/http/tests/security/window-defineProperty-crossOrigin-expected.txt

Test that calling Object.defineProperty() on a cross-origin Window throws.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.defineProperty(frames[0], 'foo', { value: 1 }) threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match..
PASS successfullyParsed is true

TEST COMPLETE
